Show offline placeholder when data is not cached
While offline, views whose data was never cached rendered nothing (blank
page) or misleading messages ("No entries available.", "Nothing found.").
Add an OfflinePlaceholder (GenericEmptyPlaceholder style) and show it when
offline with nothing to display:

- OverviewList: covers journals, entries, go-to, search and scrap views;
  rendered inside the list host so renderBeforeList stays visible
- JournalViewPage: instead of "No entries available." when entries are
  not cached
- JournalDetails: instead of a blank page when the journal itself is not
  cached
- Journals/Entries: skip "Nothing found. Try again." while offline and
  fall through to the placeholder

Extracts useIsOffline from OfflineIndicator into a shared hook.
