+++ app/src/components/overview/entries/Entries.tsx
@@ -3,18 +3,26 @@ import { useAllEntriesQuery } from "../../../serverApi/reactQuery/queries/useAll
 import { IEntry } from "../../../serverApi/IEntry";
 import { usePageContext } from "../../layout/pages/PageContext";
 import { NoResultsFound } from "../../common/search/NoResultsFound";
+import { OfflinePlaceholder } from "../../common/search/OfflinePlaceholder";
+import { useIsOffline } from "../../common/useIsOffline";
 import { OverviewList } from "../overviewList/OverviewList";
 import { EntryListItem } from "./EntryListItem";
 
 export const Entries: React.FC = () => {
   const { searchText, journalTypes } = usePageContext();
+  const isOffline = useIsOffline();
   const queryResult = useAllEntriesQuery(searchText, journalTypes);
 
   if (!queryResult) {
-    return null;
+    // undefined means not loaded (yet) - when offline that is because the
+    // query is paused until the network is back.
+    return isOffline ? <OfflinePlaceholder /> : null;
   }
 
-  if (!queryResult.entries.length && searchText) {
+  // While offline an empty result means the data is not cached (queries are
+  // paused), not that nothing matches - so fall through to the OverviewList,
+  // which renders the offline placeholder.
+  if (!queryResult.entries.length && searchText && !isOffline) {
     return <NoResultsFound />;
   }
